backend/requirements.txt
fastapi
uvicorn
python-multipart
pymongo
langchain>=0.1.0
langchain-core
langchain-community
langchain-text-splitters
langchain-openai
pypdf
python-docx
tiktoken
ujson
python-dotenv
unstructured
markdown
